Fixed Bezier rotation.
The Rotate function did not work. The code included in the step got it to work, but it was stuttering very badly, I tracked the problem down to the '|0' included at the end of this.css statements on line 66/67. Removed it and now its super fluid. I have not tested or tweaked the Arc code yet.